pyproject.toml
[tool.poetry]
name = "knowledge_base_agent"
version = "0.1.0"
description = "Knowledge Base Agent using naptha-sdk"
authors = ["Gowtham Satti <sattigowtham1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
naptha-sdk = {git = "https://github.com/NapthaAI/naptha-sdk.git"}
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
